+++ Editor/AppleAuthMacosPostprocessorHelper.cs
@@ -19,29 +19,55 @@ namespace AppleAuth.Editor
         {
             if (target != BuildTarget.StandaloneOSX)
             {
-                Debug.LogError("AppleAuthMacosPostprocessorHelper: FixManagerBundleIdentifier should only be called when building for macOS");
-                return;
+                throw new Exception(GetMessage("FixManagerBundleIdentifier should only be called when building for macOS"));
             }
-
-            const string bundleIdentifierPattern = @"(\<key\>CFBundleIdentifier\<\/key\>\s*\<string\>)(com\.lupidan)(\.MacOSAppleAuthManager\<\/string\>)";
-            const string macOSAppleAuthManagerInfoPlistRelativePath = "/Contents/Plugins/MacOSAppleAuthManager.bundle/Contents/Info.plist";
 
             try
             {
-                var macosAppleAuthManagerInfoPlistPath = path + macOSAppleAuthManagerInfoPlistRelativePath;
+                var macosAppleAuthManagerInfoPlistPath = GetInfoPlistPath(path);
                 var macosAppleAuthManagerInfoPlist = File.ReadAllText(macosAppleAuthManagerInfoPlistPath);
-                var modifiedMacosAppleAuthManagerInfoPlist = Regex.Replace(
-                    macosAppleAuthManagerInfoPlist,
-                    bundleIdentifierPattern,
-                    "$1" + PlayerSettings.applicationIdentifier + "$3");
+                var regex = new Regex(@"\<key\>CFBundleIdentifier\<\/key\>\s*\<string\>(com\.lupidan)\.MacOSAppleAuthManager\<\/string\>");
+                var match = regex.Match(macosAppleAuthManagerInfoPlist);
+                if (!match.Success)
+                {
+                    throw new Exception(GetMessage("Can't locate CFBundleIdentifier in MacOSAppleAuthManager's Info.plist"));
+                }
+                
+                var modifiedMacosAppleAuthManagerInfoPlist = macosAppleAuthManagerInfoPlist
+                        .Remove(match.Groups[1].Index, match.Groups[1].Length)
+                        .Insert(match.Groups[1].Index, PlayerSettings.applicationIdentifier);
 
                 File.WriteAllText(macosAppleAuthManagerInfoPlistPath, modifiedMacosAppleAuthManagerInfoPlist);
-                Debug.Log("AppleAuthMacosPostprocessorHelper: Renamed MacOSAppleAuthManager.bundle bundle identifier from \"com.lupidan.MacOSAppleAuthManager\" -> \"" + PlayerSettings.applicationIdentifier + ".MacOSAppleAuthManager\"");
+                Debug.Log(GetMessage($"Renamed MacOSAppleAuthManager.bundle bundle identifier from \"com.lupidan.MacOSAppleAuthManager\" -> \"{PlayerSettings.applicationIdentifier}.MacOSAppleAuthManager\""));
             }
             catch (Exception exception)
             {
-                Debug.LogError("AppleAuthMacosPostprocessorHelper: Error while fixing MacOSAppleAuthManager.bundle bundle identifier :: " + exception.Message);
+                throw new Exception(GetMessage(
+                    $"Error while fixing MacOSAppleAuthManager.bundle bundle identifier :: {exception.Message}"));
             }
+        }
+
+        private static string GetMessage(string message) => $"{nameof(AppleAuthMacosPostprocessorHelper)}: {message}";
+
+        private static string GetInfoPlistPath(string path)
+        {
+            var possibleInfoPlistPaths = new[]
+            {
+                Path.Combine(path, "Contents", "Plugins", "MacOSAppleAuthManager.bundle", "Contents", "Info.plist"),
+                Path.Combine(path, "Contents", "PlugIns", "MacOSAppleAuthManager.bundle", "Contents", "Info.plist"),
+                Path.Combine($"{path}.app", "Contents", "Plugins", "MacOSAppleAuthManager.bundle", "Contents", "Info.plist"),
+                Path.Combine($"{path}.app", "Contents", "PlugIns", "MacOSAppleAuthManager.bundle", "Contents", "Info.plist"),
+            };
+
+            foreach (var possibleInfoPlistPath in possibleInfoPlistPaths)
+            {
+                if (File.Exists(possibleInfoPlistPath))
+                {
+                    return possibleInfoPlistPath;
+                }
+            }
+            
+            throw new Exception(GetMessage("Can't locate MacOSAppleAuthManager.bundle info plist"));
         }
     }
 }